Refactor Newsfeed Posts into Posts component.
