+++ src/components/Header.js
@@ -1,4 +1,3 @@
-import React from "react";
 
 const Header = (props) => {
     return (
