+++ src/pages/About.js
@@ -2,8 +2,8 @@ import React from "react";
 
 const About = () => {
     return <div>
-            <h2>About Page</h2>
-        </div>;
+            <p>Thank you for using Cascade, a social media
+ site built using the MERN stack. This site was built as a collaboration between Alex West, Patrick Clary, and Victoria Onyema as a mid term project during a General Assembly Software Immersive bootcamp.</p>        </div>;
     };
 
 export default About;
